Fix MSVC build issues for DLL and static library
- Fix narrowing conversion warning (C4244) in CoNgramModel.cpp by adding
  explicit static_cast<uint32_t> for historyMap.size()
- Add KIWI_STATIC macro support in DllApiExport.h to make KIWI_API empty
  for static library builds (fixes C2491 dllimport definition errors)
- Define KIWI_STATIC for kiwi_static and kiwi_mt_static targets in CMakeLists.txt

+++ src/CoNgramModel.cpp
@@ -1020,7 +1020,7 @@ namespace kiwi
 						const auto historyToken = prevStates[i].history[j];
 						if (!historyToken) continue;
 						const auto idx = i * windowSize + j;
-						auto inserted = tls.historyMap.emplace(historyToken, tls.historyMap.size());
+						auto inserted = tls.historyMap.emplace(historyToken, static_cast<uint32_t>(tls.historyMap.size()));
 						tls.inverseHistoryIdcs[idx] = inserted.first->second;
 						if (inserted.second) tls.uniqHistoryTokens.emplace_back(historyToken);
 					}
